+++ CHANGELOG.md
@@ -204,6 +204,7 @@
 #### New Features
 - Added support for `DataFrame.query` for dataframes with single-level indexes.
 - Added support for `DataFrameGroupby.__len__` and `SeriesGroupBy.__len__`.
+- Added support for `Dataframe.groupby.rolling()`.
 
 #### Improvements